Refactor secrets handler and move k8s logic to k8s pkg
**What**
- Refactor the structure of the secrets handler so that it exposes a
  standard http.Handler interface for the secrets endpoints.
- Move the `UpdateSecrets` method into the k8s packages as
  `ConfigureSecrets` on the `FunctionFactory`.  This better reflects
  the intended behavior and usage of the method.  It also better
  reflects its nature as part of the k8s data layer and not the http
  layer, where it previously was.
- Refactor the kubernetes secrets logic into a SecretClient interface in
  the k8s package. This allows a clean separation of the http handler in
  the provider and the data layer in k8s. The interface allows for
  simpler testing of error cases (these new cases are still TODO)
- Refactor getlookupNamespace into NamespaceResolver so that the logic
  can be generalized and reused consistently amoung all http endpoints.
  The existing getLookupNamespace tests were refactored as minimally as
  possible to demonstrate that the new implementation is fully
  compatible with the existing implementation.
- Move and make public the ProcessErrorReasons method.  This is now in
  an errors.go file and reflects that is can be part of any http
  handler's error handling.
- Some small changes in deploy/update to use the new k8s SecretClient
  and Factory methods.
