+++ handlers/secrets_test.go
@@ -4,305 +4,282 @@
 package handlers
 
 import (
-	"errors"
+	"encoding/json"
 	"fmt"
+	"io/ioutil"
 	"net/http"
+	"net/http/httptest"
+	"strings"
 	"testing"
 
 	types "github.com/openfaas/faas-provider/types"
-	appsv1 "k8s.io/api/apps/v1"
-	apiv1 "k8s.io/api/core/v1"
-	k8serrors "k8s.io/apimachinery/pkg/api/errors"
+	v1 "k8s.io/api/core/v1"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
-	"k8s.io/apimachinery/pkg/runtime/schema"
+	testclient "k8s.io/client-go/kubernetes/fake"
 )
 
-func Test_UpdateSecrets_DoesNotAddVolumeIfRequestSecretsIsNil(t *testing.T) {
-	request := types.FunctionDeployment{
-		Service: "testfunc",
-		Secrets: nil,
-	}
-	existingSecrets := map[string]*apiv1.Secret{
-		"pullsecret": {Type: apiv1.SecretTypeDockercfg},
-		"testsecret": {Type: apiv1.SecretTypeOpaque, Data: map[string][]byte{"filename": []byte("contents")}},
-	}
+const secretLabel = "app.kubernetes.io/managed-by"
+const secretLabelValue = "openfaas"
 
-	deployment := &appsv1.Deployment{
-		Spec: appsv1.DeploymentSpec{
-			Template: apiv1.PodTemplateSpec{
-				Spec: apiv1.PodSpec{
-					Containers: []apiv1.Container{
-						{Name: "testfunc", Image: "alpine:latest"},
-					},
-				},
-			},
-		},
-	}
-	err := UpdateSecrets(request, deployment, existingSecrets)
-	if err != nil {
-		t.Errorf("unexpected error %s", err.Error())
-	}
+func Test_SecretsHandler(t *testing.T) {
+	namespace := "of-fnc"
+	kube := testclient.NewSimpleClientset()
+	secretsHandler := MakeSecretHandler(namespace, kube).ServeHTTP
+	secretName := "testsecret"
 
-	validateEmptySecretVolumesAndMounts(t, deployment)
+	t.Run("create managed secrets", func(t *testing.T) {
+		secretValue := "testsecretvalue"
+		payload := fmt.Sprintf(`{"name": "%s", "value": "%s"}`, secretName, secretValue)
+		req := httptest.NewRequest("POST", "http://example.com/foo", strings.NewReader(payload))
+		w := httptest.NewRecorder()
 
+		secretsHandler(w, req)
+
+		resp := w.Result()
+		if resp.StatusCode != http.StatusAccepted {
+			t.Errorf("want status code '%d', got '%d'", http.StatusAccepted, resp.StatusCode)
+		}
+
+		actualSecret, err := kube.CoreV1().Secrets(namespace).Get("testsecret", metav1.GetOptions{})
+		if err != nil {
+			t.Errorf("error validting secret: %s", err)
+		}
+
+		if actualSecret.Name != "testsecret" {
+			t.Errorf("want secret with name: 'testsecret', got: '%s'", actualSecret.Name)
+		}
+
+		managedBy := actualSecret.Labels[secretLabel]
+		if managedBy != secretLabelValue {
+			t.Errorf("want secret to be managed by '%s', got: '%s'", secretLabelValue, managedBy)
+		}
+
+		actualValue := actualSecret.StringData[secretName]
+		if actualValue != secretValue {
+			t.Errorf("want secret value: '%s', got: '%s'", secretValue, actualValue)
+		}
+
+		// Attempt to re-create the secret and make sure that "409 CONFLICT" is returned.
+		newReq := httptest.NewRequest("POST", "http://example.com/foo", strings.NewReader(payload))
+		newW := httptest.NewRecorder()
+		secretsHandler(newW, newReq)
+		newResp := newW.Result()
+		if newResp.StatusCode != http.StatusConflict {
+			t.Errorf("want status code '%d', got '%d'", http.StatusConflict, newResp.StatusCode)
+		}
+	})
+
+	t.Run("update managed secrets", func(t *testing.T) {
+		newSecretValue := "newtestsecretvalue"
+		payload := fmt.Sprintf(`{"name": "%s", "value": "%s"}`, secretName, newSecretValue)
+		req := httptest.NewRequest("PUT", "http://example.com/foo", strings.NewReader(payload))
+		w := httptest.NewRecorder()
+
+		secretsHandler(w, req)
+
+		resp := w.Result()
+		if resp.StatusCode != http.StatusAccepted {
+			t.Errorf("want status code '%d', got '%d'", http.StatusAccepted, resp.StatusCode)
+		}
+
+		actualSecret, err := kube.CoreV1().Secrets(namespace).Get("testsecret", metav1.GetOptions{})
+		if err != nil {
+			t.Errorf("error validting secret: %s", err)
+		}
+
+		if actualSecret.Name != "testsecret" {
+			t.Errorf("want secret with name: 'testsecret', got: '%s'", actualSecret.Name)
+		}
+
+		managedBy := actualSecret.Labels[secretLabel]
+		if managedBy != secretLabelValue {
+			t.Errorf("want secret to be managed by '%s', got: '%s'", secretLabelValue, managedBy)
+		}
+
+		actualValue := actualSecret.StringData[secretName]
+		if actualValue != newSecretValue {
+			t.Errorf("want secret value: '%s', got: '%s'", newSecretValue, actualValue)
+		}
+	})
+
+	t.Run("update non-existing secrets returns 404", func(t *testing.T) {
+		newSecretValue := "newtestsecretvalue"
+		secretName := "testsecret-doesnotexist"
+		payload := fmt.Sprintf(`{"name": "%s", "value": "%s"}`, secretName, newSecretValue)
+		req := httptest.NewRequest("PUT", "http://example.com/foo", strings.NewReader(payload))
+		w := httptest.NewRecorder()
+
+		secretsHandler(w, req)
+
+		resp := w.Result()
+		if resp.StatusCode != http.StatusNotFound {
+			t.Errorf("expected status code '%d', got '%d'", http.StatusNotFound, resp.StatusCode)
+		}
+	})
+
+	t.Run("list managed secrets only", func(t *testing.T) {
+		req := httptest.NewRequest("GET", "http://example.com/foo", nil)
+		w := httptest.NewRecorder()
+
+		secretsHandler(w, req)
+
+		resp := w.Result()
+		if resp.StatusCode != http.StatusOK {
+			t.Errorf("want status code '%d', got '%d'", http.StatusOK, resp.StatusCode)
+		}
+
+		decoder := json.NewDecoder(resp.Body)
+
+		secretList := []types.Secret{}
+		err := decoder.Decode(&secretList)
+		if err != nil {
+			t.Error(err)
+		}
+
+		secretWant := 1
+		if len(secretList) != secretWant {
+			t.Errorf("want %d secret, got %d", secretWant, len(secretList))
+		}
+
+		actualSecret := secretList[0]
+		if actualSecret.Name != secretName {
+			t.Errorf("want secret name: '%s', got: '%s'", secretName, actualSecret.Name)
+		}
+	})
+
+	t.Run("delete managed secrets", func(t *testing.T) {
+		secretName := "testsecret"
+		payload := fmt.Sprintf(`{"name": "%s"}`, secretName)
+		req := httptest.NewRequest("DELETE", "http://example.com/foo", strings.NewReader(payload))
+		w := httptest.NewRecorder()
+
+		secretsHandler(w, req)
+
+		resp := w.Result()
+		if resp.StatusCode != http.StatusAccepted {
+			t.Errorf("want status code '%d', got '%d'", http.StatusAccepted, resp.StatusCode)
+		}
+
+		actualSecret, err := kube.CoreV1().Secrets(namespace).Get("testsecret", metav1.GetOptions{})
+		if err == nil {
+			t.Errorf("want not found error, got secret payload '%s'", actualSecret)
+		}
+	})
+
+	t.Run("delete missing secret returns 404", func(t *testing.T) {
+		secretName := "testsecret-doesnotexist"
+		payload := fmt.Sprintf(`{"name": "%s"}`, secretName)
+		req := httptest.NewRequest("DELETE", "http://example.com/foo", strings.NewReader(payload))
+		w := httptest.NewRecorder()
+
+		secretsHandler(w, req)
+
+		resp := w.Result()
+		if resp.StatusCode != http.StatusNotFound {
+			t.Errorf("expected status code '%d', got '%d'", http.StatusNotFound, resp.StatusCode)
+		}
+	})
 }
 
-func Test_UpdateSecrets_DoesNotAddVolumeIfRequestSecretsIsEmpty(t *testing.T) {
-	request := types.FunctionDeployment{
-		Service: "testfunc",
-		Secrets: []string{},
-	}
-	existingSecrets := map[string]*apiv1.Secret{
-		"pullsecret": {Type: apiv1.SecretTypeDockercfg},
-		"testsecret": {Type: apiv1.SecretTypeOpaque, Data: map[string][]byte{"filename": []byte("contents")}},
-	}
+func Test_SecretsHandler_ListEmpty(t *testing.T) {
+	namespace := "of-fnc"
+	kube := testclient.NewSimpleClientset()
+	secretsHandler := MakeSecretHandler(namespace, kube).ServeHTTP
 
-	deployment := &appsv1.Deployment{
-		Spec: appsv1.DeploymentSpec{
-			Template: apiv1.PodTemplateSpec{
-				Spec: apiv1.PodSpec{
-					Containers: []apiv1.Container{
-						{Name: "testfunc", Image: "alpine:latest"},
-					},
-				},
-			},
-		},
-	}
-	err := UpdateSecrets(request, deployment, existingSecrets)
-	if err != nil {
-		t.Errorf("unexpected error %s", err.Error())
-	}
+	req := httptest.NewRequest("GET", "http://example.com/foo", nil)
+	w := httptest.NewRecorder()
 
-	validateEmptySecretVolumesAndMounts(t, deployment)
+	secretsHandler(w, req)
 
-}
+	resp := w.Result()
+	if resp.StatusCode != http.StatusOK {
+		t.Errorf("want status code '%d', got '%d'", http.StatusOK, resp.StatusCode)
+	}
+	body, _ := ioutil.ReadAll(resp.Body)
 
-func Test_UpdateSecrets_RemovesAllCopiesOfExitingSecretsVolumes(t *testing.T) {
-	volumeName := "testfunc-projected-secrets"
-	request := types.FunctionDeployment{
-		Service: "testfunc",
-		Secrets: []string{},
-	}
-	existingSecrets := map[string]*apiv1.Secret{
-		"pullsecret": {Type: apiv1.SecretTypeDockercfg},
-		"testsecret": {Type: apiv1.SecretTypeOpaque, Data: map[string][]byte{"filename": []byte("contents")}},
-	}
-
-	deployment := &appsv1.Deployment{
-		Spec: appsv1.DeploymentSpec{
-			Template: apiv1.PodTemplateSpec{
-				Spec: apiv1.PodSpec{
-					Containers: []apiv1.Container{
-						{
-							Name:  "testfunc",
-							Image: "alpine:latest",
-							VolumeMounts: []apiv1.VolumeMount{
-								{
-									Name: volumeName,
-								},
-								{
-									Name: volumeName,
-								},
-							},
-						},
-					},
-					Volumes: []apiv1.Volume{
-						{
-							Name: volumeName,
-						},
-						{
-							Name: volumeName,
-						},
-					},
-				},
-			},
-		},
-	}
-	err := UpdateSecrets(request, deployment, existingSecrets)
-	if err != nil {
-		t.Errorf("unexpected error %s", err.Error())
-	}
-
-	validateEmptySecretVolumesAndMounts(t, deployment)
-
-}
-
-func Test_UpdateSecrets_AddNewSecretVolume(t *testing.T) {
-	request := types.FunctionDeployment{
-		Service: "testfunc",
-		Secrets: []string{"pullsecret", "testsecret"},
-	}
-	existingSecrets := map[string]*apiv1.Secret{
-		"pullsecret": {Type: apiv1.SecretTypeDockercfg},
-		"testsecret": {Type: apiv1.SecretTypeOpaque, Data: map[string][]byte{"filename": []byte("contents")}},
-	}
-
-	deployment := &appsv1.Deployment{
-		Spec: appsv1.DeploymentSpec{
-			Template: apiv1.PodTemplateSpec{
-				Spec: apiv1.PodSpec{
-					Containers: []apiv1.Container{
-						{Name: "testfunc", Image: "alpine:latest"},
-					},
-				},
-			},
-		},
-	}
-	err := UpdateSecrets(request, deployment, existingSecrets)
-	if err != nil {
-		t.Errorf("unexpected error %s", err.Error())
-	}
-
-	validateNewSecretVolumesAndMounts(t, deployment)
-
-}
-
-func Test_UpdateSecrets_ReplacesPreviousSecretMountWithNewMount(t *testing.T) {
-	request := types.FunctionDeployment{
-		Service: "testfunc",
-		Secrets: []string{"pullsecret", "testsecret"},
-	}
-	existingSecrets := map[string]*apiv1.Secret{
-		"pullsecret": {Type: apiv1.SecretTypeDockercfg},
-		"testsecret": {Type: apiv1.SecretTypeOpaque, Data: map[string][]byte{"filename": []byte("contents")}},
-	}
-
-	deployment := &appsv1.Deployment{
-		Spec: appsv1.DeploymentSpec{
-			Template: apiv1.PodTemplateSpec{
-				Spec: apiv1.PodSpec{
-					Containers: []apiv1.Container{
-						{Name: "testfunc", Image: "alpine:latest"},
-					},
-				},
-			},
-		},
-	}
-	err := UpdateSecrets(request, deployment, existingSecrets)
-	if err != nil {
-		t.Errorf("unexpected error %s", err.Error())
-	}
-	// mimic the deployment already existing and deployed with the same secrets by running
-	// UpdateSecrets twice, the first run represents the original deployment, the second run represents
-	// retrieving the deployment from the k8s api and applying the update to it
-	err = UpdateSecrets(request, deployment, existingSecrets)
-	if err != nil {
-		t.Errorf("unexpected error %s", err.Error())
-	}
-
-	validateNewSecretVolumesAndMounts(t, deployment)
-
-}
-
-func Test_UpdateSecrets_RemovesSecretsVolumeIfRequestSecretsIsEmptyOrNil(t *testing.T) {
-	request := types.FunctionDeployment{
-		Service: "testfunc",
-		Secrets: []string{"pullsecret", "testsecret"},
-	}
-	existingSecrets := map[string]*apiv1.Secret{
-		"pullsecret": {Type: apiv1.SecretTypeDockercfg},
-		"testsecret": {Type: apiv1.SecretTypeOpaque, Data: map[string][]byte{"filename": []byte("contents")}},
-	}
-
-	deployment := &appsv1.Deployment{
-		Spec: appsv1.DeploymentSpec{
-			Template: apiv1.PodTemplateSpec{
-				Spec: apiv1.PodSpec{
-					Containers: []apiv1.Container{
-						{Name: "testfunc", Image: "alpine:latest"},
-					},
-				},
-			},
-		},
-	}
-	err := UpdateSecrets(request, deployment, existingSecrets)
-	if err != nil {
-		t.Errorf("unexpected error %s", err.Error())
-	}
-
-	validateNewSecretVolumesAndMounts(t, deployment)
-
-	request = types.FunctionDeployment{
-		Service: "testfunc",
-		Secrets: []string{},
-	}
-	err = UpdateSecrets(request, deployment, existingSecrets)
-	if err != nil {
-		t.Errorf("unexpected error %s", err.Error())
-	}
-
-	validateEmptySecretVolumesAndMounts(t, deployment)
-}
-
-func Test_DefaultErrorReason(t *testing.T) {
-	err := errors.New("A default error mapping")
-	status, reason := processErrorReasons(err)
-	if status != http.StatusInternalServerError {
-		t.Errorf("Unexpected default status code: %d", status)
-	}
-	if reason != metav1.StatusReasonInternalError {
-		t.Errorf("Unexpected default reason: %s", reason)
+	if string(body) == "null" {
+		t.Errorf(`want empty list to be valid json i.e. "[]", but was %q`, string(body))
 	}
 }
 
-func Test_ConflictErrorReason(t *testing.T) {
-	gv := schema.GroupResource{Group: "testing", Resource: "testing"}
-	statusError := k8serrors.NewConflict(gv, "conflict_test", errors.New("A test error for confict"))
+func Test_NamespaceResolver(t *testing.T) {
+	defaultNamespace := "openfaas-fn"
+	kube := testclient.NewSimpleClientset()
+	getLookupNamespace := NewNamespaceResolver(defaultNamespace, kube)
 
-	status, reason := processErrorReasons(statusError)
-	if status != http.StatusConflict {
-		t.Errorf("Unexpected default status code: %d", status)
-	}
-	if !k8serrors.IsConflict(statusError) {
-		t.Errorf("Unexpected default reason: %s", reason)
-	}
-}
+	allowedAnnotation := map[string]string{"openfaas": "true"}
+	allowedNamespace := "allowed-ns"
+	ns := &v1.Namespace{ObjectMeta: metav1.ObjectMeta{Name: allowedNamespace, Annotations: allowedAnnotation}}
+	kube.CoreV1().Namespaces().Create(ns)
 
-func validateEmptySecretVolumesAndMounts(t *testing.T, deployment *appsv1.Deployment) {
-	numVolumes := len(deployment.Spec.Template.Spec.Volumes)
-	if numVolumes != 0 {
-		fmt.Printf("%+v", deployment.Spec.Template.Spec.Volumes)
-		t.Errorf("Incorrect number of volumes: expected 0, got %d", numVolumes)
-	}
+	t.Run("no custom namespace", func(t *testing.T) {
+		req := httptest.NewRequest("GET", "http://example.com/foo", nil)
+		ns, err := getLookupNamespace(req)
 
-	c := deployment.Spec.Template.Spec.Containers[0]
-	numVolumeMounts := len(c.VolumeMounts)
-	if numVolumeMounts != 0 {
-		t.Errorf("Incorrect number of volumes mounts: expected 0, got %d", numVolumeMounts)
-	}
-}
+		if ns != defaultNamespace {
+			t.Errorf("expected %s, got %s", defaultNamespace, ns)
+		}
 
-func validateNewSecretVolumesAndMounts(t *testing.T, deployment *appsv1.Deployment) {
-	numVolumes := len(deployment.Spec.Template.Spec.Volumes)
-	if numVolumes != 1 {
-		t.Errorf("Incorrect number of volumes: expected 1, got %d", numVolumes)
-	}
+		if err != nil {
+			t.Errorf("expected %v, got %v", nil, err)
+		}
+	})
 
-	volume := deployment.Spec.Template.Spec.Volumes[0]
-	if volume.Name != "testfunc-projected-secrets" {
-		t.Errorf("Incorrect volume name: expected \"testfunc-projected-secrets\", got \"%s\"", volume.Name)
-	}
+	t.Run("allowed namespace in GET query parameters", func(t *testing.T) {
+		url := fmt.Sprintf("http://example.com/foo?namespace=%s", allowedNamespace)
 
-	if volume.VolumeSource.Projected == nil {
-		t.Error("Secrets volume is not a projected volume type")
-	}
+		req := httptest.NewRequest("GET", url, nil)
+		ns, err := getLookupNamespace(req)
 
-	if volume.VolumeSource.Projected.Sources[0].Secret.Items[0].Key != "filename" {
-		t.Error("Project secret not constructed correctly")
-	}
+		if ns != allowedNamespace {
+			t.Errorf("expected %s, got %s", allowedNamespace, ns)
+		}
 
-	c := deployment.Spec.Template.Spec.Containers[0]
-	numVolumeMounts := len(c.VolumeMounts)
-	if numVolumeMounts != 1 {
-		t.Errorf("Incorrect number of volumes mounts: expected 1, got %d", numVolumeMounts)
-	}
+		if err != nil {
+			t.Errorf("expected %v, got %v", nil, err)
+		}
+	})
 
-	mount := c.VolumeMounts[0]
-	if mount.Name != "testfunc-projected-secrets" {
-		t.Errorf("Incorrect volume mounts: expected \"testfunc-projected-secrets\", got \"%s\"", mount.Name)
-	}
+	t.Run("forbidden namespace in GET query parameters", func(t *testing.T) {
+		forbiddenNamespace := "forbidden"
+		url := fmt.Sprintf("http://example.com/foo?namespace=%s", forbiddenNamespace)
 
-	if mount.MountPath != secretsMountPath {
-		t.Errorf("Incorrect volume mount path: expected \"%s\", got \"%s\"", secretsMountPath, mount.MountPath)
-	}
+		req := httptest.NewRequest("GET", url, nil)
+		_, err := getLookupNamespace(req)
+
+		errMsg := "unable to manage secrets within the forbidden namespace"
+		if err.Error() != errMsg {
+			t.Errorf("expected %s, got %s", errMsg, err)
+		}
+	})
+
+	t.Run("allowed namespace in POST body", func(t *testing.T) {
+		url := "http://example.com/foo"
+		payload := fmt.Sprintf(`{"name": "secret", "value": "value", "namespace": "%s"}`, allowedNamespace)
+
+		req := httptest.NewRequest("POST", url, strings.NewReader(payload))
+		ns, err := getLookupNamespace(req)
+
+		if ns != allowedNamespace {
+			t.Errorf("expected %s, got %s", allowedNamespace, ns)
+		}
+
+		if err != nil {
+			t.Errorf("expected %v, got %v", nil, err)
+		}
+	})
+
+	t.Run("forbidden namespace in POST body", func(t *testing.T) {
+		forbiddenNamespace := "forbidden"
+		url := "http://example.com/foo"
+		payload := fmt.Sprintf(`{"name": "secret", "value": "value", "namespace": "%s"}`, forbiddenNamespace)
+
+		req := httptest.NewRequest("POST", url, strings.NewReader(payload))
+		_, err := getLookupNamespace(req)
+
+		errMsg := "unable to manage secrets within the forbidden namespace"
+		if err.Error() != errMsg {
+			t.Errorf("expected %s, got %s", errMsg, err)
+		}
+	})
 }
