+++ handlers/update.go
@@ -4,6 +4,7 @@
 package handlers
 
 import (
+	"context"
 	"encoding/json"
 	"fmt"
 	"io/ioutil"
@@ -46,7 +47,7 @@ func MakeUpdateHandler(defaultNamespace string, factory k8s.FunctionFactory) htt
 
 		annotations := buildAnnotations(request)
 		if err, status := updateDeploymentSpec(lookupNamespace, factory, request, annotations); err != nil {
-			if !isNotFound(err) {
+			if !k8s.IsNotFound(err) {
 				log.Printf("error updating deployment: %s.%s, error: %s\n", request.Service, lookupNamespace, err)
 
 				return
@@ -58,7 +59,7 @@ func MakeUpdateHandler(defaultNamespace string, factory k8s.FunctionFactory) htt
 		}
 
 		if err, status := updateService(lookupNamespace, factory, request, annotations); err != nil {
-			if !isNotFound(err) {
+			if !k8s.IsNotFound(err) {
 				log.Printf("error updating service: %s.%s, error: %s\n", request.Service, lookupNamespace, err)
 			}
 
@@ -142,12 +143,13 @@ func updateDeploymentSpec(
 
 		deployment.Spec.Template.Spec.ServiceAccountName = serviceAccount
 
-		existingSecrets, err := getSecrets(factory.Client, functionNamespace, request.Secrets)
+		secrets := k8s.NewSecretsClient(factory.Client)
+		existingSecrets, err := secrets.GetSecrets(context.TODO(), functionNamespace, request.Secrets)
 		if err != nil {
 			return err, http.StatusBadRequest
 		}
 
-		err = UpdateSecrets(request, deployment, existingSecrets)
+		err = factory.ConfigureSecrets(request, deployment, existingSecrets)
 		if err != nil {
 			log.Println(err)
 			return err, http.StatusBadRequest
